testing/rp2350_c_hid/tools/requirements.txt
pyserial
hid
